Moved decimal conversion function to utility class
